fix any bug + clean code

+++ src/controllers/users.controller.ts
@@ -6,6 +6,7 @@ import { ErrorCodes } from '~/constants/errorCodes'
 import { ParamsDictionary } from 'express-serve-static-core'
 import { USERS_MESSAGES } from '~/constants/messages'
 import RefreshTokenModel from '~/models/schemas/RefreshToken.schema'
+import { AuthenticationError } from '~/utils/CustomErrors'
 
 export const registerController = async (req: Request<ParamsDictionary, any, registerReqBody>, res: Response) => {
   const result = await registerService(req.body)
@@ -29,7 +30,7 @@ export const loginController = async (req: Request<ParamsDictionary, any, loginR
   const response = new ApiSuccess(
     ErrorCodes.SUCCESS,
     USERS_MESSAGES.LOGIN_SUCCESS,
-    201,
+    200,
     result,
     new Date().toISOString()
   )
@@ -38,6 +39,18 @@ export const loginController = async (req: Request<ParamsDictionary, any, loginR
 
 export const logoutController = async (req: Request<ParamsDictionary, any, logoutReqBody>, res: Response) => {
   const { refresh_token } = req.body
+
+  const user_id_from_access_token = (req as Request).decode_authorization?.user_id
+  const user_id_from_refresh_token = (req as Request).decode_refresh_token?.user_id
+  if(
+    !user_id_from_access_token 
+    || 
+    !user_id_from_refresh_token 
+    || 
+    user_id_from_access_token !== user_id_from_refresh_token
+  ) throw new AuthenticationError(USERS_MESSAGES.REFRESH_TOKEN_IS_INVALID)
+  
+  // bên validate đã kiểm tra và chắc chắn có rồi nên ko cần check lại nữa
   await RefreshTokenModel.deleteOne({
     token: refresh_token
   })
